mock crypter moved to test package
So the file is not included when the module is vendored in the downstream projects
